issue.1 - Moving ListDeleted method from Vaults to DeletedVaults
