+++ azure-mgmt-keyvault-preview/src/main/java/com/microsoft/azure/management/keyvault/implementation/VaultsImpl.java
@@ -13,7 +13,6 @@ import com.microsoft.azure.management.keyvault.Vaults;
 import com.microsoft.azure.management.keyvault.Vault;
 import rx.Observable;
 import rx.Completable;
-import com.microsoft.azure.management.keyvault.DeletedVault;
 import com.microsoft.azure.management.keyvault.CheckNameAvailabilityResult;
 import rx.functions.Func1;
 import java.util.ArrayList;
@@ -179,42 +178,6 @@ class VaultsImpl extends GroupableResourcesImpl<Vault, VaultImpl, VaultInner, Va
         return wrapModel(name);
     }
 
-    private Observable<Page<DeletedVaultInner>> listDeletedNextInnerPageAsync(String nextLink) {
-        if (nextLink == null) {
-            Observable.empty();
-        }
-        VaultsInner client = this.inner();
-        return client.listDeletedNextAsync(nextLink)
-        .flatMap(new Func1<Page<DeletedVaultInner>, Observable<Page<DeletedVaultInner>>>() {
-            @Override
-            public Observable<Page<DeletedVaultInner>> call(Page<DeletedVaultInner> page) {
-                return Observable.just(page).concatWith(listDeletedNextInnerPageAsync(page.nextPageLink()));
-            }
-        });
-    }
-    @Override
-    public Observable<DeletedVault> listDeletedAsync() {
-        VaultsInner client = this.inner();
-        return client.listDeletedAsync()
-        .flatMap(new Func1<Page<DeletedVaultInner>, Observable<Page<DeletedVaultInner>>>() {
-            @Override
-            public Observable<Page<DeletedVaultInner>> call(Page<DeletedVaultInner> page) {
-                return listDeletedNextInnerPageAsync(page.nextPageLink());
-            }
-        })
-        .flatMapIterable(new Func1<Page<DeletedVaultInner>, Iterable<DeletedVaultInner>>() {
-            @Override
-            public Iterable<DeletedVaultInner> call(Page<DeletedVaultInner> page) {
-                return page.items();
-            }
-       })
-        .map(new Func1<DeletedVaultInner, DeletedVault>() {
-            @Override
-            public DeletedVault call(DeletedVaultInner inner) {
-                return new DeletedVaultImpl(inner);
-            }
-       });
-    }
     @Override
     public Observable<CheckNameAvailabilityResult> checkNameAvailabilityAsync(String name) {
         VaultsInner client = this.inner();
